fix: move NATS message definitions into separate package
This change fixes a bug where ssh-portal binaries were unnecessarily
exposing ssh-portal-api related metrics (these metrics are always zero
in ssh-portal).

Previously ssh-portal's sshserver package was importing sshportalapi
(and thus getting the unnecessary metrics) just for the bus message
struct definitions.

By moving the message struct definitions out into their own package,
both sshserver and sshportalapi can import the message definitions as
required without any side effects (such as metrics registrations).
